+++ Yang.Weather.DataAccess/Models/GeoCode.cs
@@ -3,8 +3,8 @@
     public class GeoCode
     {
         public int Id { get; set; }
-        public float Latitude { get; set; }   
-        public float Longitude { get; set; }
+        public decimal Latitude { get; set; }   
+        public decimal Longitude { get; set; }
         public string Postalcode { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
     }
